Enable mapping on a SMRTLink server, via pbservice
This gets mapping working via SMRTLink.  On successful mapping on
SMRTLink, we then link to the final alignmentset.xml result produced
by SMRTLink, and our own workflow can continue.  We also create a
symlink to the job directory to enable easy inspection in the future.

There are a few imperfections here:

- At the moment, I am hardcoding the path the pbsmrtpipe preset XML.
  This needs to be bundled as a resource.

- I am using the pbservice command and parsing its output.  This is
  really not the ideal approach--we should be able to just use the
  REST services API as surfaced in Python in pbcommand.  This is on
  the docket.

- I should output the job ID and URI to a file for easy pickup by Zia

- We should enable Zia to customize the name of SMRTLink jobs so they
  can appear something like "ZiaJob345-ConditionNameX" on the SMRTLink
  server.