Add UnrolledNoHQMappingReports workflow
This commit adds similar plots as in MappingReports workflow to UnrolledNoHQMappingReports. The plots should be double checked to make sure they are good for unrolled data.
In this commit, the unrolled mapping data are nolonger saved to folder "conditions/{condition}/mapped_unrolledNoHQ/", cause all plotting scripts are using "conditions/{condition}/mapped/" to input alignments. Since we are using the same R scripts to generating plots for MappingReports and UnrolledMappingReports, I believe it is easier to leave the alignments in one folder.