Enable bauhaus2 to load sts.h5 file and make relative plots
This commit enables bauhaus2 to load sts.h5 files at the same time when reference is loaded. When sts.h5 exists, a link to the sts.h5 will be created under the condition folder. When sts.h5 file does not exist, a link points to an empty sts.h5 file will be created. If non-empty sts.files exist for all conditions within a job, several plots with productivity and read type information will be generated.