Increase the sample size for hmm model
Base on Alicia's request, the sample size for each condition in the constant arrow model is increased to 1000 from 500. 500 was suggested by Enzymology group since they have less ZMWs for each reference. Currently, if the number of ZMWs for on condition is less than 1000, all ZMWs will be used to make the fishbone plots.