Load reference from Smrtlink Jobs
Foe mapped alignments from Smrtlink, the referece should be directly loaded from the smrtlink job. Previously the reference is loaded from the input genome name, and may cause confilictions when the genome name is inconsistent with the reference used in the smrtlink job. In this commit, when the input is mapped alignments from smrtlink jobs, the reference link will be created to the real reference used in the smrtlink job.