Use xpath to load the reference from the alignmentset.xml
This commits uses xpath to load the reference from the alignmentset.xml. Note that the xpath command here selects the "ResourceId" attribute and output 'ResourceId="xxx.fasta"'. The reference path is then separated by an awk command. It seems that xpath2.0 can output only the value of the attribute by string(), but our xpath module is 1.13, and using string() will generate an error "Query didn't return a nodeset".