Change `constant_arrow.R` to have the CLI option to sample ZMWs for different references
Now the script constant_arrow.R can load an option from CLI to decide whether the ZMWs are sampled for different references. The default is set to "FALSE" so current snakemake code still works. When diferent references are used for enzymology group, the command in the snakemake file can be changed to "Rscript --vanilla scripts/R/constant_arrow.R --sampleByRef TRUE".

This commit is made for enzymology group so that they can rewrite the condition table (chunked by reference) to run constant arrow.