Redo Mapping use current Genome for Coverage Titration
This function requires us to collect the subreads from smrtlink jobs. The most convenient way would be load the subreadset from the mapped alignmentset. However, the original subreadset is not saved as a tag in current smrtlink mapped alignmentsets, so in this commit, the way bauhaus2 resolves smrtlink jobs was revised to load the subreadset as well as the alignmentset from the pbalign jobs. In coverage titration workflow, even if mapped alignments are used as input, we still redo the mapping use the provided Genome.