Add comments to expalin the resolved data structure of smrtlink jobs.
at the upstream resolver, when resolve the smrtlink jobs, we would like to get both the alignmentsets and the subreadsets. The subreadsets are now added into the resolver and in the condition table, for any smrtlink job, the alignmentset and the subreadset are now return as a list in ct.inputs(c). In the following snakemake workflow, to make sure a correct data structure, we define the list into "remote_alignmentsets" and then use different entries to get the alignmentset or the subreadset.