Revert to the most recent working version
This branch would like to add maron I to the heatmaps. The newest version of the script from Onureena does not work. So this commit revert it to the most recent version. The reason I'm using this temprary version is that I need to check bamboo build on the branch.