Workaround to get the job id from pbservice json output
There was an design problem in peservice to output json files. Basiclly, the log of the job is also saved in the json output before the "real" json output. the commit adds a gross workaround to read the job is and path from the incorrect json output. When SE-660 is solved, this commit shoyld be revert back.